refactor: 도메인 계층의 dto를 도메인으로 내부로 이동
