+++ src/main/java/roomescape/application/dto/response/reservation/UserReservationResponse.java
@@ -4,7 +4,7 @@ import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import java.time.LocalDate;
 import java.time.LocalTime;
-import roomescape.domain.dto.ReservationWithRank;
+import roomescape.domain.reservation.ReservationWithRank;
 import roomescape.domain.reservation.Reservation;
 
 public record UserReservationResponse(
